requirements.txt
Django==4.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn
pytest-django
pytest